fix: ignore broad exceptions on raise keyword
If a user uses a broad exception handler, but raises the exception again via `raise`, we should not raise an error.

This change isn't perfect.

The following two examples should both fail, but we don't raise an error for them. We'd need to track the control flow which is a pain.
```
# alpha
try:
    print("Hello")
except (ValueError, BaseException):
    if x != 123:
        return
    raise

# beta
for x in my_results:
    try:
        print("Hello")
    except (ValueError, BaseException):
        if x != 123:
            if y == 10:
                continue
        raise
```