ref: combine visitors
By using a checker for each rule we were iterating over the AST each time.
So for 10 rules we'd iterate over the same module 10 times, not great.

Instead we combine the logic into one visitor. There's probably a better
way to abstract this, but for now this matches the behavior and offers
better perf.